chore(filters): move ClickableLabel to separate source file
